Comment out package declaration
When the package declaration is wrong (e.g. `BitSet.java` isn't in the
folder `java/io/`), then in Visual Studio Code the Java plugin won't
show certain errors or warnings (example: "The value of the field ... is
not used").

Instead of moving the file, I've commented out the package declaration.
This shouldn't make a difference for KeY.