Make BitSet loadable in KeY
Changes:
- Comment out imports
- Internally override OpenJDK library methods
- Comment out methods which use OpenJDK library methods which are too
  cumbersome to override
- Add comments to class members which aren't used anywhere due to
  commented out code

KeY cannot load BitSet without changes, because it doesn't recognize all
OpenJDK classes and methods. To fix this, we comment out any imports and
internally (in the class) declare missing dependencies ourselves. This
has been done for some library classes which are essential.

We don't just declare these library methods - we also include an
implementation that we can verify later with KeY when we're verifying
BitSet as a whole. This is also why some library methods that KeY knows
of are overriden here anyway (e.g. `Math.max` and `Math.min`) because
they don't have a JML contract in KeY.

For OpenJDK class methods which are too cumbersome to override and add
an implementation for (often due to transitive dependencies), we've
commented out any of BitSet's methods which use these classes. For
instance, all "Won't-have" methods are commented out entirely.

After these changes, KeY will load the file without error.