Add intger set sort definition with axiomatic taclets
Adds a ".key" file with definitions for an integer set sort definition
for use with a ghost variable in BitSet. For technical reasons, the sort
is called "Free" instead of IntSort, or similar.

The file "project.key" is necessary for loading the "BitSet.java" file
in KeY together with the taclet file.

This new sort is not yet used in BitSet's JML.