Remove internal fields from public JML contracts
"External users" of the BitSet class with these JML contracts have no
use for ensures or requires clauses which reference internal fields. The
object invariants should be sufficient anyways to enforce a consistent
internal state (unless public methods within BitSet call each other and
rely upon additional knowledge of the internal workings of methods -
which is not the case with the "Must-have" methods).

Updated proof files.