Verify `get` and extract bitwise operation
In order to successfully verify `get(int)` we extract the bitwise
operation on `words[]` into a seperate helper function.

We also need the precondition that all elements of `words[]` are
exclusively in the range of a `long`. This should be implicit in KeY,
but for some reason at the moment this won't verify without the
precondition.

To verify any method that uses `checkInvariants()` it is important to
remember to use the Use Dependency Contract rule manually and select the
`normal_behavior` contract. Otherwise KeY will merge the two contracts
for the method and create a new contract which overapproximates the
`assignable` clause for the two contracts.