Move condition `inLong(words[i])` from `get(int)` into object invariants
It is always true (given `wellFormed(heap)`) that the elements of the
(non-null) `words` array of BitSet are of type `long` and
`inLong(words[i])` is true for all valid indices `i`. KeY cannot prove
this (see KeY issue #1271), thus it was added to `get(int)`'s contract
preconditions to make it possible to prove the contract.

The condition is now moved to the object invariants because it seems
like the `length()` method would also benefit from it. Also, we don't
want to reference private fields in public preconditions of methods, so
it was misplaced in `get(int)`'s contract anyway. Having these private
fields exposed also makes it harder to use the specification of a method
in other KeY projects, since you need to first prove that the
precondition is true - but it references a private field which may not
be in reach.